"Increased border widths for section elements and added about section with typing animation"
This commit includes changes to the CSS file and the index.html file. In the CSS file, the border widths for section elements have been increased from 1px to 2px. Additionally, a new about section has been added to the index.html file, which includes a typing animation for the heading.