"Updated service icons in index.html"
This commit message describes the changes made in the code diff for the index.html file. The service icons have been updated with new images, specifically "img/pfc2.png" for Passing Funded Challenges, "img/rm2.png" for Risk Management, and "img/sd2.png" for Strategy Development. The message is concise and accurately reflects the modifications in the code.