Added contact section with Twitter link
This commit adds a new section to the index.html file for contact information, which includes a link to the author's Twitter profile. The contact section is positioned at the bottom of the page and is implemented using Bootstrap grid system. The Twitter link is styled using Font Awesome icons.